Motivation Scenario - AEC
This is an example of a heterogeneous document corpus in the context of the Architecture, Engineering and Construction (AEC) industry, i.e.,  describing a construction project and involving contracts, technical specification documents, on-site photos, 2D drawings, etc. exchanged between architects, owners, contractors and consultants throughout the whole life-cycle of the project.